Re-add items for UH-1H and A-10C that were removed in v0.8.4
This commit renames the v0.7.1 exports for the Huey's radar altimeter
displays, chaff and flare count displays to add a "_STR" suffix.

That provides compatibility with hardware cockpits (as the address
is the same as the v0.7.1 address) and frees up the name for the new
additions so users of the DCSFlightpanels fork should not be affected
either, as DCSFlightpanels references the outputs by name instead
of address.